Commented out errors, Party System
Commented out lines 46, 47, and 60 in playerMovement.cs as it was causing compilation errors and unreadable console feedback. Commented out line 37 in Crouch.cs as it was causing compilation errors. Commented out all of climbing in Dash.cs as it was causing compilation errors. Commented out the update function in combatHandler.cs as it was causing errors. Fixed and implemented fully functioning data system for player and enemy party inside combatHandler.cs