Fixed CombatHandler
Combat Handler now works fine with public constructors, values of party are editable through inspector. Line 46 and 47 of movement is commented out for compiling, same with all of climbing the in the Dash script. Camera lag code was fixed but was commented out as it was not working as intended.